refactor: AdminController 패키지 구조 & 네이밍 변경 리팩토링
